requirements.txt
codeclimate-test-reporter==0.1.1
coverage==4.1
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.6
django-floppyforms==1.6.2
django-ordered-model==1.1.0
django-smart-selects==1.2.2
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.1.1
poirot==0.2.0
psycopg2==2.6.1
py==1.4.31
PyJWT==1.4.0
pytest==2.9.2
pytest-cov==2.2.1
pytest-django==2.9.1
python-social-auth==0.2.19
python3-openid==3.0.10
regex==2016.5.23
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
tqdm==4.7.1
whitenoise==3.2
